Changes to local_xmlsync and core for Moodle upgrade.
WR396437

This is a combination of 4 commits, it involves the below files:

       modified:   backup/externallib.php
       modified:   backup/util/helper/copy_helper.class.php
       modified:   enrol/database/lib.php
       modified:   lib/classes/task/asynchronous_copy_task.php
       new file:   local/xmlsync/classes/privacy/provider.php
       modified:   local/xmlsync/classes/util.php
       modified:   local/xmlsync/lang/en/local_xmlsync.php

-------------------
For reference the details from the original commits are mentioned below
-------------------
This is the 1st commit message:

Changes made to local_xmlsync and core_backup to support the switch from using deprecated copy.php to copy_helper class.

WR396437

modified:   backup/externallib.php
modified:   backup/util/helper/copy_helper.class.php
new file:   local/xmlsync/classes/privacy/provider.php
modified:   local/xmlsync/classes/util.php
modified:   local/xmlsync/lang/en/local_xmlsync.php

This is the commit message #2:

Core change to enrol_database: clone courses templated from {local_xmlsync_crsimport} records.

Creates:
    \local_xmlsync\util::enrol_database_template_check()
    \local_xmlsync\util::enrol_database_template_hook()
    \local_xmlsync\util::enrol_database_course_hook()
    \local_xmlsync\util::enrol_database_course_update_hook()

If the course_idnumber matches and the course_template field is a
valid course, create a copy.

Copies are not immediate; they are triggered via the sync_courses()
method and will be completed on a cron run.

This also checks visibility of the course from the import system
and sets/updates it

WR#371793

This is the commit message #3:

enrol_database Fix hooks for xmlsync

This is the commit message #4:

Add support for resetting time columns on course copy

For the use case unitec has, they don't want to reset the timemodified
and coursecreated values to the same as the target copy course, so we
just set them to the time the task runs